Updated Post testing
Took away seaborn imports (don't exist on controls PC's)
Updated the equation for the Decaying Cosine
Contemplated but did not implement decaying sine vs. cosine 
Added in a plotting section that can be used for testing to make sure that you actually reach zero
Updated the plotting axes to display time instead of arbitrary units. Will not be updating in the IPYNB