Created Puck
This could be used for the future to make increasingly complex tasks more simple. This acts as a way to create virtual objects in the magnet settings space. In the distant future, these pucks would be created for each adjacent magnet configuration. Or there would be one puck that includes all of the different magnet settings.